Fix bugs in HDA, PCI and syscall subsystems
HDA driver:
- Decrement num_codecs and free widget memory on parse failure
- Add spinlock protection to shared stream state

PCI driver:
- Fix signed integer overflow (1UL << 31) in legacy config access
- Fix pci_update_usable_list discarding finding results
- Fix pci_device_find memory leak on unknown type
- Fix pci_legacy_write corrupting adjacent bytes on unaligned access
- Fix base_address_register_t address type truncating 64-bit MMIO

Syscall:
- Fix set_tid_address truncating user pointer to 32 bits
- Add clear_child_tid zero write on process exit
- Rewrite select() to use proper fd_set bitmask semantics
- Set -EINTR on syscall return after signal delivery
- Fix fchdir setting basename instead of full path
- Relax SYS_CLONE validation to allow valid stack pointer
- Add clock_settime, chmod, fchmod, chown, fchown, mknod implementations
- Add realtime_base_ns for clock_gettime/gettimeofday/time

+++ include/proc/process.h
@@ -118,7 +118,7 @@ typedef struct process {
         char              name[PROCESS_NAME_LEN];
         char              root[256];       /* chroot path */
         char              cwd[256];        /* current working directory */
-        int32_t           clear_child_tid; /* set_tid_address target */
+        uint64_t          clear_child_tid; /* set_tid_address target */
 } process_t;
 
 /* Initialize the process management subsystem */
